requirements.txt
anyio==4.0.0
certifi==2023.7.22
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
pip==23.3.1
python-telegram-bot==20.6
setuptools==68.2.2
sniffio==1.3.0
wheel==0.41.3

boto3~=1.29.3
python-dotenv~=1.0.0